+++ crypto/src/basics/mod.rs
@@ -1,8 +1,8 @@
 pub mod chaum_pedersen;
 pub mod elgamal;
 pub mod hybrid_encryption;
+pub mod matrix_sigma;
 pub mod pedersen_elgamal;
 pub mod rescue;
 pub mod ristretto_pedersen_comm;
 pub mod schnorr;
-pub mod sigma;
