+++ crypto/src/basics/pedersen_elgamal.rs
@@ -1,6 +1,6 @@
 use crate::basics::elgamal::{ElGamalCiphertext, ElGamalEncKey};
+use crate::basics::matrix_sigma::{sigma_prove, sigma_verify_scalars, SigmaProof, SigmaTranscript};
 use crate::basics::ristretto_pedersen_comm::RistrettoPedersenCommitment;
-use crate::basics::sigma::{sigma_prove, sigma_verify_scalars, SigmaProof, SigmaTranscript};
 use curve25519_dalek::traits::{Identity, MultiscalarMul};
 use merlin::Transcript;
 use zei_algebra::prelude::*;
